requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
av==15.1.0
certifi==2026.2.25
cffi==2.0.0
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.7.1
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
humanfriendly==10.0
idna==3.11
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==26.0
protobuf==6.33.5
pycparser==2.23
Pygments==2.19.2
PySide6_Essentials==6.10.2
PyYAML==6.0.3
rich==14.3.3
scipy==1.13.1
shellingham==1.5.4
shiboken6==6.10.2
sounddevice==0.5.5
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.2
typing_extensions==4.15.0
